Refactor commit handling in control plane deployment scripts to improve error handling and streamline push logic

+++ deploy/scripts/pipeline_scripts/01-control-plane-deploy.sh
@@ -397,16 +397,17 @@ git pull -q origin "$BRANCH"
 if [ 1 = $added ]; then
   git config --global user.email "$BUILD_REQUESTEDFOREMAIL"
   git config --global user.name "$BUILD_REQUESTEDFOR"
-  git commit -m "Added updates from Control Plane Deployment for $DEPLOYER_FOLDERNAME $LIBRARY_FOLDERNAME $BUILD_BUILDNUMBER [skip ci]"
-  if git -c http.extraheader="AUTHORIZATION: bearer $SYSTEM_ACCESSTOKEN" push --set-upstream origin "$BRANCH" --force-with-lease ; then
-    echo "##vso[task.logissue type=error]Failed to push changes to the repository."
+  if git commit -m "Added updates from Control Plane Deployment for $DEPLOYER_FOLDERNAME $LIBRARY_FOLDERNAME $BUILD_BUILDNUMBER [skip ci]"; then
+    if git -c http.extraheader="AUTHORIZATION: bearer $SYSTEM_ACCESSTOKEN" push --set-upstream origin "$BRANCH" --force-with-lease; then
+      echo "##vso[task.logissue type=error]Failed to push changes to the repository."
+    fi
   fi
 
 fi
 
-if [ -f ".sap_deployment_automation/${ENVIRONMENT}${LOCATION}.md" ]; then
-  echo "##vso[task.uploadsummary].sap_deployment_automation/${ENVIRONMENT}${LOCATION}.md"
-fi
+# if [ -f ".sap_deployment_automation/${ENVIRONMENT}${LOCATION}.md" ]; then
+#   echo "##vso[task.uploadsummary].sap_deployment_automation/${ENVIRONMENT}${LOCATION}.md"
+# fi
 
 if [ -f "${deployer_environment_file_name}" ]; then
